1st Place/requirements.txt
absl-py==1.0.0
aiosignal==1.2.0
attrs==21.4.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
cycler==0.11.0
distlib==0.3.4
filelock==3.6.0
fonttools==4.33.3
frozenlist==1.3.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
grpcio==1.43.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
joblib==1.1.0
jsonschema==4.4.0
kiwisolver==1.4.2
Markdown==3.3.6
matplotlib==3.5.1
msgpack==1.0.3
numpy==1.22.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
pkg_resources==0.0.0
platformdirs==2.5.2
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybaselines==0.8.0
pyparsing==3.0.8
pyrsistent==0.18.1
python-dateutil==2.8.2
pytorch-model-summary==0.1.2
pytz==2022.1
PyYAML==6.0
ray==1.12.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
sklearn==0.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
timm==0.5.4
torch==1.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
virtualenv==20.14.1
Werkzeug==2.1.2
zipp==3.8.0
